refactor: move types and interface
