+++ src/lib/AbstractValidationWather.ts
@@ -1,31 +1,9 @@
-export interface Error {
-  id: string;
-  message: string;
-}
-
-export interface Validator<T, Action = any> {
-  error: Error;
-  afterReduce?: boolean;
-  idSelecter?(id: number, action: Action): number;
-  idSelecter?(id: string, action: Action): string;
-  validate(state: T, action?: Action): boolean;
-}
+import { Error } from "./types";
 
 export type ReturnType = "object" | "array";
-
 export type ResultValue = Error[] | { [id: string]: Error };
-
-export interface Config<T extends string, R extends ReturnType> {
-  errorStateId?: T;
-  returnType?: R;
-}
-
 export interface InternalParams<T extends ResultValue> {
   results: T;
-}
-
-export interface Reducer<State extends {}, Action> {
-  (state: State, action: Action): State;
 }
 
 export abstract class AbstractValidationWatcher<
